Send back target path for easier using navigo
When we catch a navigate event, and we want to do ajax call based on the catched url, and we use just some general route like /user/:action/:id in our rooter, then we need to know what was the originally catched url, without rebuild the original url from response parameters, query strings or without any hacky other solutions). I know this is not the most elegant way, but it doesn't break anything and better then nothing. In v5 navigo should send back an object with { params, query, targetPath } instead of 3 arguments.